Composer: one auto-grow hook, in platform
The hero composer grew with its content by writing an inline pixel
height off scrollHeight — which goes stale the moment the box's WIDTH
changes and the same text rewraps to more lines. After a sidebar toggle
or a window resize the box kept its old height and quietly turned into a
scroller, or held a tall empty gap. The starter briefs are the longest
text that lands in there, so the chips made it easy to hit.

The Playground's `useAutoGrow` already had this fixed, with the failure
mode written out in its doc comment: a ResizeObserver watching the width
only (reacting to height would loop on what the callback itself does).
It moves to `platform/lib/autoGrow.ts` — two apps need it now, and an
app may only import platform + itself, so a shared copy is the only way
the composer gets the fix rather than a second hand-rolled version of
the same bug. `max` gains the `Infinity` case for a box whose ceiling is
a CSS max-height in rows rather than a pixel count, which is what the
composer's 3-to-10-row band is.

The composer drives it from an effect on `prompt` rather than onChange,
because a starter chip sets the text programmatically.
