+++ frontend/src/apps/ai_models/playground/TextStage.tsx
@@ -46,9 +46,9 @@ import {
   ResultSlot,
   StageHeader,
   StarterCards,
-  useAutoGrow,
   type Starter,
 } from "./controls";
+import { useAutoGrow } from "@platform/lib/autoGrow";
 import { StarterIcons } from "./starterIcons";
 import { numParam, readParam, writeParams } from "@apps/ai_models/lib/params";
